requirements.txt
djangorestframework
django
django-rest-auth